requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages

apache-airflow-providers-google
python-dotenv
google-cloud-container
google-cloud-storage
google-cloud-bigquery
pandas